requirements.txt
altgraph==0.17
colorama==0.4.4
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
pefile==2021.5.24
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
smmap==4.0.0